+++ src/assembler_arm.h
@@ -481,8 +481,7 @@ class ArmAssembler : public Assembler {
   virtual void CopyRef(FrameOffset dest, FrameOffset src,
                        ManagedRegister scratch);
 
-  virtual void Copy(FrameOffset dest, FrameOffset src, ManagedRegister scratch,
-                    unsigned int size);
+  virtual void Copy(FrameOffset dest, FrameOffset src, ManagedRegister scratch, size_t size);
 
   virtual void Copy(FrameOffset dest, ManagedRegister src_base, Offset src_offset,
                     ManagedRegister scratch, size_t size);
